+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### 2026-04-05
+- C++ enum definitions added (enums.h): CreatureSize, ItemType, ItemSource, ItemRarity, WeaponCategory, WeaponRangeType, DamageType, ArmorCategory
+
 ### 2026-03-28
 - SQLite schema added: 6 tables (characters, character_coins, item_definitions, weapon_details, armor_details, inventory_items)